raster: check the provided buffer size correctly
We mustn't rely on cap(p.line) because we're using append, and it may
round up the capacity to a page boundary, even when the appended data
fit exactly.

Instead, just check against the known byte count per line.

Additionally, use 3-index slicing in ReadAll to make sure that buggy code doesn't
access parts of the buffer it isn't supposed to.